Clarify wording in make-request
This is a very nitpicky pull request, so I apologize for that. I was/am using this as a model for my own REST-wrapper of the Twilio API and felt badly about not contributing back to this; since I can't assess or really suggest improvements to the code, I thought clearing up the wording in the documentation would be the least I could do.